% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.geom_massity.r
\docType{data}
\name{geom_massity}
\alias{geom_massity}
\alias{GeomMassity}
\alias{stat_massity}
\alias{StatMassity}
\title{Densité non normalisée}
\usage{
geom_massity(
  mapping = NULL,
  data = NULL,
  stat = "massity",
  position = "identity",
  ...,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "upper"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{outline.type}{encadrement des rectangles ("both", "upper", "lower", "full")}

\item{geom}{The geometric object to use display the data}
}
\value{
une "layer" pour ggplot
}
\description{
Affiche une fonction de densité, discrétisée, exprimée en unité de y de sorte que la surface est égale à la somme de y sur l'espace des x
procure quelques fonction accessibles par after_stat (density, mass, cummass, totalmass, mean)
permet de comparer des "densités" entre elles puisqu'elles sont exprimées en unité commune (et non pas normées à 1 en surface)
}
\seealso{
[geom_massogram()] densité en masse, en transformant l'échelle des x pour avoir une distribution par quantile,
  [geom_quantogram()] calcule la densité et normalise les valeurs de x pour que chaque intervalle de x soit un quantile de la distribution de x,
  [geom_quantagram()] calcule la densité et normalise les valeurs de x mais conserve l'échelle initiale de x
}
\keyword{datasets}
